+++ Apps/System/tests/vibes_test.c
@@ -5,7 +5,7 @@
  * Author: Author: Elliot Hawkins <elliotshawkins@gmail.com>.
  */
 
-#include "rebbleos.h"
+#include "librebble.h"
 #include "systemapp.h"
 #include "menu.h"
 #include "status_bar_layer.h"
