+++ lib/png/png.h
@@ -2,8 +2,7 @@
 
 #include <stdint.h>
 #include <stdbool.h>
-#include <pebble.h>
+#include "gbitmap/gbitmap.h"
 
-
-void png_to_gbitmap(GBitmap *bitmap, uint8_t *raw_buffer, size_t png_size);
+void png_to_gbitmap(n_GBitmap *bitmap, uint8_t *raw_buffer, size_t png_size);
