+++ rcore/api_func_symbols.h
@@ -3,6 +3,7 @@
 #include "librebble.h"
 #include "graphics_wrapper.h"
 #include "battery_state_service.h"
+#include "librebble.h"
 
 GBitmap *gbitmap_create_with_resource_proxy(uint32_t resource_id);
 bool persist_exists(void);
