+++ rcore/appmanager.c
@@ -10,13 +10,14 @@
 #include <stdlib.h>
 #include "rebbleos.h"
 #include "librebble.h"
-#include "appmanager.h"
 #include "systemapp.h"
 #include "test.h"
 #include "notification.h"
 #include "api_func_symbols.h"
 #include "qalloc.h"
 #include "notification_manager.h"
+
+#include "appmanager.h"
 
 /* Configure Logging */
 #define MODULE_NAME "appman"
