+++ rwatch/ui/animation/animation.h
@@ -8,7 +8,7 @@
 
 #include "FreeRTOS.h"
 
-#include "librebble.h"
+//#include "librebble.h"
 #include "point.h"
 #include "rect.h"
 #include "size.h"
